requirements.txt
# Core dependencies
streamlit
langchain
langchain_ollama
langchain_groq
langchain-google-genai

# Vector store and embeddings
chromadb

# Document processing
beautifulsoup4
requests

# Environment and utilities
python-dotenv
uuid



# Additional dependencies
langchain_community
chroma
google-generativeai